# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/work-orders": {
      "post": {
        "tags": [
          "work-orders"
        ],
        "summary": "Post Work Order",
        "operationId": "post_work_order_api_v1_work_orders_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Work Orders",
        "operationId": "get_work_orders_api_v1_work_orders_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "work_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Type"
            }
          },
          {
            "name": "assigned_to_current_user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Assigned To Current User"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Get Audit Journal",
        "description": "List journaled user actions, newest first.\n\nArgs:\n    user_id: Optional exact actor filter.\n    action: Optional exact action filter, e.g. \"POST /api/v1/work-orders\".\n    target_type: Optional exact target resource type filter.\n    target_id: Optional exact target resource identifier filter.\n    from_: Optional ISO 8601 inclusive lower bound on occurred_at.\n    to: Optional ISO 8601 inclusive upper bound on occurred_at.\n    cursor: Opaque pagination cursor from the previous page's meta.\n    limit: Page size, 1-200.\n    _: The caller; must hold audit.read.\n\nReturns:\n    An AuditListEnvelope with one page of entries.",
        "operationId": "get_audit_journal_api_v1_audit_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessGrantOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "work_order_id": {
            "type": "string",
            "title": "Work Order Id"
          },
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "access_level": {
            "type": "string",
            "const": "technical_full",
            "title": "Access Level",
            "default": "technical_full"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "offline_cache_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offline Cache Expires At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "work_order_id",
          "facility_id",
          "user_id",
          "starts_at",
          "expires_at",
          "revoked_at",
          "status",
          "offline_cache_expires_at",
          "version"
        ],
        "title": "AccessGrantOut"
      },
...
      "AuditListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/AuditListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "AuditListEnvelope",
        "description": "The standard {data, meta} envelope for GET /api/v1/audit."
      },
      "AuditListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "AuditListMeta",
        "description": "Cursor-pagination metadata for the audit journal envelope."
      },
...
      "CreateWorkOrderRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "source_risk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Risk Id"
          },
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "target_entity_type": {
            "type": "string",
            "title": "Target Entity Type"
          },
          "target_entity_id": {
            "type": "string",
            "title": "Target Entity Id"
          },
          "work_type": {
            "type": "string",
            "title": "Work Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "symptoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symptoms"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "client_occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Occurred At"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "facility_id",
          "target_entity_type",
          "target_entity_id",
          "work_type",
          "priority",
          "due_at",
          "description"
        ],
        "title": "CreateWorkOrderRequest",
        "description": "Backward-compatible draft creation request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepairResultOut": {
        "properties": {
          "failure_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Confirmed"
          },
          "root_cause_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Code"
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/PartUsage"
            },
            "type": "array",
            "title": "Parts"
          },
          "labor_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Labor Minutes"
          },
          "equipment_restored": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Restored"
          },
          "control_check_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Check Result"
          },
          "residual_risk": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Residual Risk"
          },
          "recommendations": {
            "type": "string",
            "title": "Recommendations"
          },
          "requires_follow_up": {
            "type": "boolean",
            "title": "Requires Follow Up"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRefOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "failure_confirmed",
          "root_cause_code",
          "diagnosis",
          "actions",
          "parts",
          "equipment_restored",
          "control_check_result",
          "residual_risk",
          "recommendations",
          "requires_follow_up",
          "completed_at",
          "author"
        ],
        "title": "RepairResultOut"
      },
...
      "SlaOut": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "acceptance_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Due At"
          },
          "arrival_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Due At"
          },
          "resolution_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Due At"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "paused_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused At"
          },
          "pause_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Reason"
          },
          "breach_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breach Stage"
          },
          "remaining_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Seconds"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "started_at",
          "acceptance_due_at",
          "arrival_due_at",
          "resolution_due_at",
          "current_stage",
          "state",
          "paused_at",
          "pause_reason",
          "breach_stage",
          "remaining_seconds",
          "server_time"
        ],
        "title": "SlaOut"
      },
...
      "UserRefOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "title": "UserRefOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkOrderAssignmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "work_order_id": {
            "type": "string",
            "title": "Work Order Id"
          },
          "engineer_id": {
            "type": "string",
            "title": "Engineer Id"
          },
          "assigned_by": {
            "$ref": "#/components/schemas/UserRefOut"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "declined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declined At"
          },
          "decline_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decline Reason"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "work_order_id",
          "engineer_id",
          "assigned_by",
          "assigned_at",
          "accepted_at",
          "declined_at",
          "decline_reason",
          "completed_at",
          "status",
          "version"
        ],
        "title": "WorkOrderAssignmentOut"
      },
      "WorkOrderListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/WorkOrderListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "WorkOrderListEnvelope"
      },
      "WorkOrderListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "WorkOrderListMeta"
      },
      "WorkOrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_number": {
            "type": "string",
            "title": "Display Number"
          },
          "source_risk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Risk Id"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          },
          "target_entity_type": {
            "type": "string",
            "title": "Target Entity Type"
          },
          "target_entity_id": {
            "type": "string",
            "title": "Target Entity Id"
          },
          "work_type": {
            "type": "string",
            "title": "Work Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "symptoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symptoms"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "assigned_engineer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Engineer Id"
          },
          "coordinator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinator Id"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "sla_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Policy Id"
          },
          "sla": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SlaOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "repair_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepairResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_assignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkOrderAssignmentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "access_grant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccessGrantOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_number",
          "source_risk_id",
          "facility_id",
          "target_entity_type",
          "target_entity_id",
          "work_type",
          "priority",
          "due_at",
          "description",
          "comment",
          "status",
          "created_by",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "WorkOrderOut",
        "description": "One work order plus its current assignment, grant, SLA, and actions."
      },
...
    }
  }
}